refactor: split endpoint definitions to their own file
